Update README.md
I have no direct interest in writing a book over these discoveries to come, but I was merely interested.  I haven't worked on it much since I wrote it many, many months ago, but it functions as it should!  Still, there may be updates to give.  I also started an iPhone App equivalent in XCode, so if I ever finish that, or a version in Python, I will share that as well!